refactor(core): split bind and scale modules; extract run setup
Break bind into bind-data/bind-layer and scale-training into
scale-axis/scale-color facades. Extract setupPipelineRun so pipeline.ts
is a short phase chain (~250 lines). Tests cover setup normalize/flip
and theme rejection contracts.
